Move layout to separate folder
