+++ utmap-client/src/components/layout/Footer.js
@@ -4,7 +4,7 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Avatar from '@material-ui/core/Avatar';
 import Grid from '@material-ui/core/Grid';
-import avatars from './data/Contributors'
+import avatars from '../data/Contributors'
 
 const useStyles = makeStyles(theme => ({
     divider: {
